ццццццц

+++ Serhii_C/Django_Sergiy_C/myapp/views.py
@@ -5,9 +5,12 @@ from .forms import NoteForm
 from .models import Note
 from .models import Category
 from .forms import CategoryForm
+from django.contrib.auth.decorators import login_required,permission_required
 
+@login_required
 def note_list(request):
     notes = Note.objects.all().order_by('published_date')
+    # notes = Note.objects.filter(user=request.user).order_by('published_date')
 
     search_query = request.GET.get('q')
     if search_query:
@@ -35,15 +38,20 @@ def note_list(request):
     }
 
     return render(request, 'myapp/note/note_list.html', context)
-
+@login_required
 def note_detail(request, pk):
     note = get_object_or_404(Note, pk=pk)
-    context = {'note': note}
-    return render(request, 'myapp/note/note_detail.html', context)
+    if request.user == note.user:
+        context = {'note': note}
+        return render(request, 'myapp/note/note_detail.html', context)
+    else:
+
+        return redirect('list')
 
 
 
-
+@login_required
+@permission_required('is_staff')
 def note_create(request):
     if request.method == 'POST':
         form = NoteForm(request.POST)
@@ -60,26 +68,36 @@ def note_create(request):
     elif request.method == 'GET':
         form = NoteForm()
         return render(request, 'myapp/note/note_form.html', {'form': form})
-
+@login_required
 def note_update(request, pk):
     note = get_object_or_404(Note, pk=pk)
-    if request.method == 'POST':
-        form = NoteForm(request.POST,instance=note)
-        if form.is_valid():
-            note = form.save(commit=False)
-            note.save()
-            return redirect('detail',pk=note.pk)
+    if request.user == note.user:
+        if request.method == 'POST':
+            form = NoteForm(request.POST,instance=note)
+            if form.is_valid():
+                note = form.save(commit=False)
+                note.save()
+                return redirect('detail',pk=note.pk)
+        else:
+            form = NoteForm(instance=note)
+        return render(request, 'myapp/note/note_form.html', {'form': form, 'note': note})
     else:
-        form = NoteForm(instance=note)
-    return render(request, 'myapp/note/note_form.html', {'form': form, 'note': note})
+
+        return redirect('list')
+@login_required
 def note_delete(request, pk):
     note = get_object_or_404(Note, pk=pk)
-    if request.method == 'POST':
-        note.delete()
-        return redirect('list')
-    elif request.method == 'GET':
-        return render(request, 'myapp/note/note_delete.html', {'note': note})
+    if request.user == note.user:
+        if request.method == 'POST':
+            note.delete()
+            return redirect('list')
+        elif request.method == 'GET':
+            return render(request, 'myapp/note/note_delete.html', {'note': note})
+    else:
 
+        return redirect('list')
+
+@login_required
 def create_category(request):
     if request.method == "POST":
         form = CategoryForm(request.POST)
@@ -91,7 +109,7 @@ def create_category(request):
     return render(request, 'myapp/note/category_form.html', {'form': form})
 
 def search(request):
-    for notes in Note.objects.all():
+    for notes in Note.objects.filter(user=request.user).order_by('published_date'):
         if notes.title==request.GET.get('title'):
             note_detail(request,notes.pk)
             return render(request, 'myapp/note/note_detail.html', {'note': notes})
